+++ tests/xy_c_compile_resources/fors/for3.c
@@ -7,29 +7,29 @@ typedef struct for3_Array for3_Array;
 
 struct for3_Array {
     float m_elems[100];
-    size_t m_len;
+    uint64_t m_len;
 };
 
-size_t for3_iter(void) {
+uint64_t for3_iter(void) {
     return 0;
 }
 
-void for3_next(size_t* p_idx) {
+void for3_next(uint64_t* p_idx) {
     (*p_idx)++;
 }
 
-float* for3_get(for3_Array* p_arr, size_t p_idx) {
+float* for3_get(for3_Array* p_arr, uint64_t p_idx) {
     return &p_arr->m_elems[p_idx];
 }
 
-void for3_set(for3_Array* p_arr, size_t p_idx, float p_elem) {
+void for3_set(for3_Array* p_arr, uint64_t p_idx, float p_elem) {
     p_arr->m_elems[p_idx] = p_elem;
 }
 
 float for3_sum(for3_Array* p_arr) {
-    size_t tmp_0_arg = for3_iter();
+    uint64_t tmp_0_arg = for3_iter();
     float l_res = 0;
-    for (size_t tmp_1_iter = tmp_0_arg; tmp_1_iter < p_arr->m_len; for3_next(&tmp_1_iter)) {
+    for (uint64_t tmp_1_iter = tmp_0_arg; tmp_1_iter < p_arr->m_len; for3_next(&tmp_1_iter)) {
         if (!(tmp_1_iter < p_arr->m_len)) {
             abort();
         }
@@ -40,13 +40,13 @@ float for3_sum(for3_Array* p_arr) {
 }
 
 float for3_mix(for3_Array* p_arr1, for3_Array* p_arr2) {
-    size_t tmp_0_arg = for3_iter();
-    size_t tmp_2_arg = for3_iter();
+    uint64_t tmp_0_arg = for3_iter();
+    uint64_t tmp_2_arg = for3_iter();
     float l_res = 0;
     {
         uint32_t i = 0;
-        size_t tmp_1_iter = tmp_0_arg;
-        size_t tmp_3_iter = tmp_2_arg;
+        uint64_t tmp_1_iter = tmp_0_arg;
+        uint64_t tmp_3_iter = tmp_2_arg;
         for (; tmp_1_iter < p_arr1->m_len && tmp_3_iter < p_arr2->m_len; ++i, for3_next(&tmp_1_iter), for3_next(&tmp_3_iter)) {
             if (!(tmp_1_iter < p_arr1->m_len)) {
                 abort();
@@ -63,8 +63,8 @@ float for3_mix(for3_Array* p_arr1, for3_Array* p_arr2) {
 }
 
 void for3_double(for3_Array* p_arr1) {
-    size_t tmp_0_arg = for3_iter();
-    for (size_t tmp_1_iter = tmp_0_arg; tmp_1_iter < p_arr1->m_len; for3_next(&tmp_1_iter)) {
+    uint64_t tmp_0_arg = for3_iter();
+    for (uint64_t tmp_1_iter = tmp_0_arg; tmp_1_iter < p_arr1->m_len; for3_next(&tmp_1_iter)) {
         if (!(tmp_1_iter < p_arr1->m_len)) {
             abort();
         }
@@ -95,8 +95,8 @@ void for3_changeSomehow(float* p_f) {
 }
 
 void for3_iterAndChange(for3_Array* p_arr1) {
-    size_t tmp_0_arg = for3_iter();
-    for (size_t tmp_1_iter = tmp_0_arg; tmp_1_iter < p_arr1->m_len; for3_next(&tmp_1_iter)) {
+    uint64_t tmp_0_arg = for3_iter();
+    for (uint64_t tmp_1_iter = tmp_0_arg; tmp_1_iter < p_arr1->m_len; for3_next(&tmp_1_iter)) {
         if (!(tmp_1_iter < p_arr1->m_len)) {
             abort();
         }
