+++ tests/xy_c_compile_resources/strings/stringCtor.c
@@ -11,10 +11,10 @@ struct stringCtor__EMPTY_STRUCT_ {
 };
 struct stringCtor_Str {
     void* m_addr;
-    size_t m_size;
+    uint64_t m_size;
 };
 
-stringCtor_Str stringCtor_str(void* p_addr, size_t p_size) {
+stringCtor_Str stringCtor_str(void* p_addr, uint64_t p_size) {
     return (stringCtor_Str){p_addr, p_size};
 }
 
@@ -24,6 +24,6 @@ void stringCtor_createStrings(void) {
     const stringCtor_Str l_str2 = stringCtor_str((int8_t*)"def", 3);
 }
 
-size_t stringCtor_strLen(stringCtor_Str p_str) {
+uint64_t stringCtor_strLen(stringCtor_Str p_str) {
     return p_str.m_size;
 }
